requirements.txt
flask
yfinance
pandas
numpy
datetime
tensorflow-cpu
scikit-learn
keras-tuner
Flask-Cors
gunicorn